pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-segmentation-benchmark"
version = "1.0.0"
description = "A comprehensive benchmarking framework for medical image segmentation models, specifically designed for intracranial aneurysm detection and segmentation using the ADAM dataset."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@domain.com"},
]
maintainers = [
    {name = "Your Name", email = "your.email@domain.com"},
]
keywords = [
    "medical imaging",
    "segmentation",
    "deep learning",
    "benchmarking",
    "aneurysms",
    "ADAM dataset",
    "computer vision",
    "healthcare AI"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "nibabel>=5.0.0",
    "SimpleITK>=2.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-image>=0.20.0",
    "scikit-learn>=1.3.0",
    "opencv-python>=4.8.0",
    "scipy>=1.11.0",
    "psutil>=5.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.4.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/medical-segmentation-benchmark"
Documentation = "https://medical-segmentation-benchmark.readthedocs.io"
Repository = "https://github.com/your-username/medical-segmentation-benchmark.git"
Issues = "https://github.com/your-username/medical-segmentation-benchmark/issues"
Changelog = "https://github.com/your-username/medical-segmentation-benchmark/blob/main/CHANGELOG.md"

[project.scripts]
medical-benchmark = "unified_benchmark:main"
benchmark-medical = "unified_benchmark:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["data*", "models*", "training*", "evaluation*", "utils*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.md", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["data", "models", "training", "evaluation", "utils"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "nibabel.*",
    "SimpleITK.*",
    "wandb.*",
    "tensorboard.*",
    "hydra.*",
    "omegaconf.*",
    "rich.*",
    "click.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
    "data: marks tests that require data files",
]

[tool.coverage.run]
source = ["data", "models", "training", "evaluation", "utils"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", "env"]
skips = ["B101", "B601"]

[tool.safety]
output = "json"
